I've fixed a cascade of build and linting issues that were preventing your application from building successfully.
Here's a summary of the changes I made:
- Corrected the Tailwind CSS configuration for `darkMode`.
- Replaced `@apply` with direct CSS rules in `globals.css` to avoid build errors.
- Moved the NextAuth `authOptions` to a separate file to fix a Next.js App Router export issue.
- Updated an incorrect import path for `authOptions`.
- Fixed various ESLint errors and warnings, including `prefer-const`, `no-unescaped-entities`, and unused variables.
- Addressed `react-hooks/exhaustive-deps` warnings by using `useCallback` to memoize functions.
