+++ visual-algo/src/app/protected/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth/next"
-import { authOptions } from "@/app/api/auth/[...nextauth]/route"
+import { authOptions } from "@/lib/auth"
 
 export default async function ProtectedPage() {
   const session = await getServerSession(authOptions);
