+++ visual-algo/src/lib/auth.ts
@@ -1,0 +1,25 @@
+import { type NextAuthOptions } from "next-auth"
+import CredentialsProvider from "next-auth/providers/credentials"
+
+export const authOptions: NextAuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "email", placeholder: "test@example.com" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        // Add your own authentication logic here
+        // For now, we'll just accept any user with any password
+        if (credentials) {
+          return { id: "1", name: "Test User", email: credentials.email }
+        }
+        return null
+      }
+    })
+  ],
+  pages: {
+    signIn: '/auth/signin',
+  }
+};
